+++ src/test/java/com/thoughtworks/go/scm/plugin/model/requestHandlers/CheckoutRequestHandlerTest.java
@@ -4,8 +4,8 @@ import com.thoughtworks.go.plugin.api.request.GoPluginApiRequest;
 import com.thoughtworks.go.plugin.api.response.GoPluginApiResponse;
 import com.thoughtworks.go.scm.plugin.HelperFactory;
 import com.thoughtworks.go.scm.plugin.util.JsonUtils;
-import com.tw.go.plugin.GitHelper;
 import com.tw.go.plugin.cmd.ProcessOutputStreamConsumer;
+import com.tw.go.plugin.git.GitHelper;
 import com.tw.go.plugin.model.GitConfig;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -34,7 +34,7 @@ public class CheckoutRequestHandlerTest {
     @Mock
     private GitConfig gitConfigMock;
 
-    private String revision = "b6d7a9c";
+    private final String revision = "b6d7a9c";
     private final String destinationFolder = "destination";
 
     @Test
